Uniprot, plus some other changes
Former-commit-id: 04b7429700541617b8da620547f927d5d9e7e00f

+++ flymine/model/mage/src/java/org/flymine/dataconversion/MageConverter.java
@@ -52,7 +52,7 @@ import org.apache.log4j.Logger;
  */
 public class MageConverter extends FileConverter
 {
-    private int opCount;
+    private long opCount;
     private long time;
     private long start;
